requirements.txt
praw
instabot
pillow